requirements.txt
python-dotenv
infisical